requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pymongo==4.10.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-multipart==0.0.12

# OAuth
authlib==1.3.0
httpx==0.27.0
itsdangerous==2.2.0

# Environment management
python-dotenv==1.0.0

# Email validation
email-validator==2.1.1
